headers showing up on additional photos screen but now the uid has gone missing
