+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "expoServerPort": 19000,
   "expoServerNgrokUrl": "https://p6-x5s.trevorseitz.lm-native.exp.direct",
   "packagerNgrokUrl": "https://packager.p6-x5s.trevorseitz.lm-native.exp.direct",
-  "ngrokPid": 36439,
+  "ngrokPid": 25476,
   "devToolsPort": 19002,
-  "packagerPort": 19006,
-  "packagerPid": 36411
+  "packagerPort": 19008,
+  "packagerPid": 25449
 }
