+++ navigation/AppDrawerNavigator.js
@@ -11,7 +11,7 @@ import DashboardScreen from "../screens/DashboardScreen";
 import HomeScreen from "../screens/HomeScreen";
 import ListLocationsScreen from "../screens/ListLocationsScreen";
 import AddLocationScreen from "../screens/AddLocationScreen";
-import MultiLocationScreen from "../screens/MultiLocationScreen";
+// import MultiLocationScreen from "../screens/MultiLocationScreen";
 import ProfileScreen from "../screens/ProfileScreen";
 import LogoutScreen from "../screens/auth/LogoutScreen";
