+++ navigation/switches/AppSwitchNavigator.js
@@ -13,8 +13,8 @@ import SignupScreen from "../../screens/auth/SignupScreen";
 import ForgotPasswordScreen from "../../screens/auth/ForgotPasswordScreen";
 import LoadingScreen from "../../screens/auth/LoadingScreen";
 import ImageBrowser from "../../screens/ImageBrowser";
-import AdditionalPhotosBrowser from "../../screens/AdditionalPhotosBrowser";
-import MultiLocationScreen from "../../screens/MultiLocationScreen";
+import AdditionalPhotosScreen from "../../screens/AdditionalPhotosScreen";
+// import MultiLocationScreen from "../../screens/MultiLocationScreen";
 import ImageTile from "../../screens/ImageTile";
 
 import AppDrawerNavigator from "../AppDrawerNavigator";
@@ -31,9 +31,9 @@ const AppSwitchNavigator = createSwitchNavigator(
     EditLocation: { screen: EditLocationScreen },
     Welcome: { screen: WelcomeScreen },
     Profile: { screen: ProfileScreen },
-    MultiLocation: { screen: MultiLocationScreen },
+    // MultiLocation: { screen: MultiLocationScreen },
     ImageBrowser: { screen: ImageBrowser },
-    AdditionalPhotos: { screen: AdditionalPhotosBrowser },
+    AdditionalPhotos: { screen: AdditionalPhotosScreen },
     ImageTile: { screen: ImageTile },
     Loading: { screen: LoadingScreen }
   },
