+++ screens/Map.js
@@ -11,6 +11,7 @@ export default class Map extends React.Component {
     super(props);
     this.state = {
       uid: this._retrieveData(),
+      user: null,
       location: null,
       locations: [],
       checkLocation: {},
@@ -24,11 +25,60 @@ export default class Map extends React.Component {
   _retrieveData = async () => {
     try {
       const value = await AsyncStorage.getItem("uid");
+      console.log(value)
       if (value !== null) {
         this.setState({ uid: value });
       }
     } catch (error) {}
   };
+
+  _getUserAsync = async () => {
+    const { navigation } = this.props;
+    const uid = await AsyncStorage.getItem("uid");
+    let { status } = await Permissions.askAsync(Permissions.USER);
+    if (status !== "granted") {
+      this.setState({
+        errorMessage: "Permission to access user was denied"
+      });
+    }
+
+    let user = await firebase
+      .firestore()
+      .collection("users")
+      .doc(JSON.parse(navigation.getParam(uid)))
+      .get()
+      .then(doc => {
+        if (doc.exists) {
+          this.setState({
+            user: doc.data(),
+            // key: doc.id,
+            isLoading: false
+          });
+        } else {
+          console.log("No such document!");
+        }
+      });;
+    this.setState({ user });
+  };
+
+  _getUserName = async () => {
+    firebase
+      .firestore()
+      .collection("users")
+      .doc(JSON.parse(navigation.getParam(uid)))
+      .get()
+      .then(doc => {
+        if (doc.exists) {
+          this.setState({
+            user: doc.data(),
+            // key: doc.id,
+            isLoading: false
+          });
+        } else {
+          console.log("No such document!");
+        }
+      });
+  }
 
   _storeData = async user => {
     try {
@@ -138,6 +188,7 @@ export default class Map extends React.Component {
 
     return (
       <View style={styles.container}>
+        <Text>Hello {}</Text>
         <MapView
           style={styles.map}
